rename package from enforceproto to unknownproto
Addressing @aaronc's feedback on the name bikeshed.
